Add code formatting test to test script

+++ scripts/test.js
@@ -3,22 +3,26 @@ const deployContracts = require('./deploy-contracts')
 const testContracts = require('./test-contracts')
 const startIpfs = require('./start-ipfs')
 const testJavascript = require('./test-javascript')
+const testFormat = require('./test-format')
 
 const start = async () => {
+  let daemon
   try {
     await startGanache()
     await deployContracts()
     await testContracts()
-    let daemon = await startIpfs()
+    daemon = await startIpfs()
     await testJavascript()
-    console.log('All tests successful. :)')
-    daemon.kill() // otherwise this process will keep running in the background
-    process.exit()
+    await testFormat()
+    console.log('Tests passed. :)')
   }
   catch(error) {
-    console.log('Some tests failed. :(')
-    process.exit()
+    console.log(`Tests failed.\n\n${error}`)
   }
+  if (daemon) {
+    daemon.kill()
+  }
+  process.exit()
 }
 
 start()
